bugs 56262, 62620
This checkin includes the following

- iochannel framework.  It allows sending a packet to peer machines
in the cluster using TCP socket.  There are two persistent unidirectional
TCP connections maintained for each pair of the machines.  The connections
are managed using NIO Channel and Selector by one Client thread and
one Server thread.  When the connection is interrupted, iochannel will
attempt to re-establish the connection.

- MessageChannel.  MessageChannel is an instance of iochannel running
on ZCS mailbox servers.  It populates peer servers based on LDAP
Server entries with mailbox service enabled.  Then it allows the
caller application to send a message to an account on another mailbox
server.

- CrossMailboxNotification.  This is one of the app registered in
MessageChannel to take advantage of iochannel messaging framework.
It is being used by RemoteSoapChannel to push the Mailbox notification
data to an account on another mailbox host, allowing instant
Mailbox notification delivery in multi node environment.