jira: HD-478, HD-2657
changed the lifecycle of OperationContext to be longer lived.
instead of instantiating new OperationContext for each request
we cache the instance in Session so it can be reused as long
as the session is valid.  therefore any data cached in
OperationContext is also longer lived.

changed WatchCache to support changing watch mapping of
item that lives on another mailbox server.  when a watch
for item is changed, and if the auth user is not on the
same mailboxd, the server sends a message to the
auth user's mailboxd using iochannel.

fixed the bug HD-2657 where the watch activity was
still enclosed in <created/> block.

https://reviewboard.eng.vmware.com/r/337063