Bug:74148
Time difference between SendMsgRequest and SaveDraftRequest is 68 ms (comment 1).  During sendMsgRequest autosavetimer is killed, but autosavetimer is not killed in certain race condition (depends on browsers cleartimeout method). Checking private variable of autosavetimer to check whether cleartimeout is called or not before saving the draft.

This bug is not reproducible locally. I used settimeout for autosave before SendMsgRequest to reproduce this bug.

http://bugzilla.zimbra.com/show_bug.cgi?id=74148