Bug 74260
Shows error "Unknown runtime error" While applying/removing tag to any document
The tag image element (div in case on IE, img in case of others) had the same id as the containing TD. No reason for the TD to have ID at all. But I believe on FF/etc getElementById was returning the IMG tag, while for IE, the TD. The parent thus became the TR, and setting the image as the innerHTML of the TR is both illegal and wrong even if it worked.  In this case I would even say the IE behavior is reasonable (except of course the useless cryptic error message).

http://bugzilla.zimbra.com/show_bug.cgi?id=74260