bug: 74204 - Notification block doesn't returned with attribute "d" when granting a share
After change 369763, now single OperationContext instance is shared during a Session. But the GranteeNames data needs to vary depending on the context. In somes cases GranteeNames.needGranteeName is set to false, though it is true by default. Hence now using a ThreadLocal instance to hold the GranteeNames data instead of keeping it in OperationContext.

http://bugzilla.zimbra.com/show_bug.cgi?id=74204