Bug 69262
IE : Can't click to alphabet button cell area (td DwtButton AlphabetBarCell)
I tried to figure out in any logical way what IE9's problem might be, but couldn't. However, wrapping the inner text of the TD with a span with width 100% fixed it, so I guess we'll have to add that. I guess for some reason IE9 decided it's cool to only apply the TD attributes such as style and onclick, onmouseover, etc, to the text within the TD and not the entire TD. Doesn't make any sense to me.

http://bugzilla.zimbra.com/show_bug.cgi?id=69262