Bug: 74073 NO such blob exception when the server does auto savedraft.
The problem caused by change 368029 from bug 71778. The message object is uncached
and later updated and in the meantime older version of the same item is cached. The next
update writes the older value of mod_content for the item from the cache. Reverting the
change 368029; Updating recipients in Message object in reanalyze() to fix Bug 74073. 

https://reviewboard.eng.vmware.com/r/338724/

http://bugzilla.zimbra.com/show_bug.cgi?id=74073