jira: HD-478
added proxying support for fetching activities for shared
items on a different mailboxd.  RemoteEventLog makes
GetActivityStreamRequest against the remote mailboxd,
caches the result and makes it available to the local
MergedEventLog.

tweaked the merging algorithm for MergedEventLog when
prefetching more items.

added count attribute to GetActivityStreamResponse
to show how many events are in the database for
that particular item.  MergedEventLog aggregates
the total number of events across multiple logs
accordingly.

fixed typo in WatchCache that was setting the
timeout value to 10ms instead of 10s.

fixed typo in Db.getEventsBefore.

https://reviewboard.eng.vmware.com/r/339097