jira: HS-1912
Add a new <GetShareDetailsRequest> SOAP op.  It returns a trimmed version of the item grants (just perms, grantee type, grantee display name, grantee email) in the <GetShareDetailsResponse>, but the SOAP response structure should be easy to extend with other share data if needed in the future.  No grantee data is returned if you request share details on the subfolder of a shared folder (i.e. we return only the actual ACL data, not the effective ACL data).  External users are denied access to this SOAP command for security reasons.