Bug 65499
Can't sort by To in message mode with reading pane on right
fixed the problem from comment #5.
Also fixed a seemingly simple problem, that the "To" field was missing from the column actions menu of outbound folders (such as "sent") in case of viewing pane on right (single column).  This was more complicated, mostly due to taking a while to realize the calls to DwtMenu.prototype.getItemById where wrong in a few places! That function expects a key param, before the ID param... and does
not work with just ID param (it would always return the first item, which is a bug but in some cases, such as getting the From field, it worked by luck since it was first...).   The best fix is to use getMenuItem(id) which works fine (or to pass the key to getItemById(ZmOperation.MENUITEM_ID, id) but that's not necessary, and verbose.
In any case, I changed the logic for the actions menu (for the viewing pane on right case only - I didn't touch the multi-column case - too scary refactor for now) so instead of using just the From field and renaming it "From" or "To", something I don't like so much - I have both From and To (see ZmMailListView.SINGLE_COLUMN_SORT), and hide the irrelevant menu item based on whether the folder is outbound or inbound.

http://bugzilla.zimbra.com/show_bug.cgi?id=65499