bug 72369
= keyboard shortcut key for "Get Mail" doesn't work
for some unknown reason, on FF/WIN the key code (part of this.__keySequence in DwtKeyboardMgr.prototype.__dispatchKeyEvent) is not 61 ("=") or the one it is on Chrome, 187 (which is some platform specific thing). Instead of those 2 that are supported so far, it is 107 (lower case "k". no idea why). In any case since the real "K" key thing we support (for previous message) gives the uppercase ascii code when clicking (75), I think it's safe to add 107 as a possible key that means "=".

http://bugzilla.zimbra.com/show_bug.cgi?id=72369