bug 74392 improvement in the client logic in detecting whether the socket is currently connected, and if not takes itself out of the active set of peers.
deleted seq attribute in the notify block in order to
improve duplicate notification detection.

the message channel client will not attempt to connect
to local server any more.

reduced the amount of log message in case of non fatal
error when some of the servers are configured out of
the system.

http://bugzilla.zimbra.com/show_bug.cgi?id=74392