Integrate change 375342 from HELIX bug: 31536 bug: 74401
Make sure CreateMailbox operation is logged; it was not being logged since mailbox was not open yet.
Handle mailbox ID conflicts during redo replay when invoked from command line. This avoids problems which can occur during parallel playback; the order in which separate mailboxes are created may not always be the same as the original order. 

These changes allow restore from redo log as long as the LDAP is backed up and restored to a point in time where the account already existed. In most cases this means LDAP replication between the active site and the standby site. If the account does not exist in the LDAP backup, it must be re-provisioned before mailbox data can be recovered.

http://bugzilla.zimbra.com/show_bug.cgi?id=31536