Bug 74290
DL Edit page: Error message for saving an empty member list
changed the error message for DL to "Distribution List members must be
specified.". Note that the DL name check was already done separately so no need to combine (not sure why it is in the group case) the two to one message. Moved the DL name test to before the "empty" test so that the correct message is displayed in case the DL name is empty.

http://bugzilla.zimbra.com/show_bug.cgi?id=74290