Bug 74297
"Cc" field doesn't show DL address, while click on 'Reply All' link from footer
another bizarre IE bug. same bubble works fine in other places, like the message header.  But for some reason in the quick reply part, the "white-space: nowrap" seems to cause the text to disappear. Again, no idea why.
I added a special rule to override the white-space: nowrap only for the conv2 quick reply area. I don't think it would wrap in most cases anyway for this.
Made it just for IE to be on the safe side for the better browsers.
Note that it also happened in selected bubble (clicked/focused) and actioned (right clicked). So added rules for that too.

http://bugzilla.zimbra.com/show_bug.cgi?id=74297