Bug 74597
DL cannot be added to the Contact Group
treat DL like a regular member in this case, as it's not a regular group.
I commit my fix, as it's mostly independent of the server-side fix required to include the "ref" attribute in GetAccountDistributionListsResponse.
After the server-side fix I should check that it all works correctly and probably remove the || contactToAdd.id fallback from _createContactRefObj

http://bugzilla.zimbra.com/show_bug.cgi?id=74597