Bug 75115
Vertical scrollbar slightly cutdown in free/busy view in normal and full message view
One thing that was wrong was getting the size using getSize() without
getFromStyle param set to true, which gives 2 pixels off compared to the real width set in the style. If the width is set explicitly to a pixel value, as is in this case, it's better to call getSize(true).
But this only reduced 2 pixels from the cutdown.  There are still 2 more off. I tried to figure out the root cause of that (the parent calendar_view has it's width set wrongly, not sure why).
I saw one thing I thought was the cause of the 2 extra pixels off, but that
change doesn't seem to have a real effect in our case. Hopefully it doesn't cause regressions. This is the fix in DwtControl.prototype.setBounds - when specific pixel number is set to the width or height, no need to use the getBounds result for the event for
notification, since it's, again, off in 2 pixels (or other values) and the only
reason to call it is if we set something like "80%".
One last thing, so eventually I adjusted the extra 2 pixels. However, those were needed only if viewing pane is on right. But another bug was we re-use the day/col view when switching reading pane from button to right and vice versa - thus the _isRight value was wrong on switching view. So I fixed that too.

http://bugzilla.zimbra.com/show_bug.cgi?id=75115