Bug 72070
Unable to click "Previous/Next" button on "To: > Select addresses" page
the refactoring/fixing I did for bug 72008 lays the foundation to do this
easier, as it separated the concepts of server-side pagination from client-side pagination. For this I also cleaned up the code way more, reduced code duplication, etc.  The core issue with GAL (no sync) was just that the "more" condition was not calculated correctly in case more data came from the server than was requested (again since the GAL ignores the limit). But I saw so many redundancies and unnecessary complications in the code, I had to refactor and reduce them, making the code more stable and maintainable.
To make it clear, in the case of GAL without sync, if we get 140 results back, we stash them just like in other cases, and first display 50, with possibility for the user to paginate to next 50 and next 40 after that. The only difference is we can't/don't paginate on the server-side via offset, as it's not supported.
Added bonus: fixed the "no results found" case - it was not showing it (at least when switching the source (gal/contacts/emailed contacts), and searching, but I think probably in all cases. Just 2 different bugs, one the "firstTime" was not set correctly, though now I don't use it for this, and the other that it was overridden by the _chooser._setItems after it.
I tested locally by setting ZmContactsApp.SEARCHFOR_MAX to values such as 5 and 4 so I get pagination even with my 10 GAL results, etc.

http://bugzilla.zimbra.com/show_bug.cgi?id=72070