Bug 66564
Distribution List large icon should be displayed on the right panel
used the "person with scroll behind" icon for both large and small DL icon.
I copied the ImgDistributionList from ImgDistributionListGroup (both under admin) so the name is shorter. The original ImgDistributionList was the clipboard icon and Josh doesn't think it should be different.

http://bugzilla.zimbra.com/show_bug.cgi?id=66564