Bug 71403
When "Allow the user to specify a forwarding address" is OFF, then incorrectly label is shown in a config window in Ajax.
there was some fragile "if"s there to figure out when the label should be displayed.
I made it generic, calling a method from each of the optional places and only display the label the first time.

http://bugzilla.zimbra.com/show_bug.cgi?id=71403