Bug 75332
Wrong message opened after attaching mime message
fix to the case the attachment message is attached to a message with no body text (empty fragment in the top level message)
unfortunately size 0 condition doesn't work, as I believe it includes stuff
other than the body size, (probably subject, addresses, etc).
so for now, for an isRfc822 popup, I simply do not display "The message has no text content." in any case. It's a rare enough edge case and it's not really abug to show the empty body if it was sent empty. So I believe it's fine for now.

http://bugzilla.zimbra.com/show_bug.cgi?id=75332