Bug 75508
Adding Sent To address adds duplicate query bubbles
aparantly no need to have a key down handler to handle "enter" as it's already handled by the ZmAutocompleteListView.onKeyDown aclv.handleAction call that triggers down the call stack a call to _updateContinuation which runs a callback to type "comp" which is
ZmAddressSearchFilter.prototype._acCompHandler. That updates the bubbles the first time and that's enough.
An extra bonus - fixed the bug where the address input field (from the "sent to" or "received from" menues) was pre-populated after one search to the next. Clear the address input on update.
And one more extra bonus - fixed a null exception in some cases on the domain search due to a regression from bug 68966.

http://bugzilla.zimbra.com/show_bug.cgi?id=75508