bug 74240 - Issue with folders names in french accent
When issuing an AJAX request, the browser runtime can recompose
decomposed Unicode characters. In this particular instance, it
converted LATIN SMALL LETTER E, U+0065, followed by a COMBINING ACUTE
ACCENT, U+0301, to a LATIN SMALL LETTER E WITH ACUTE, U+00E9. The
server has no notion of Unicode equivalence, and thus isn't 'clever'
enough to reverse the transliteration, so it fails to find the
requested folder.

The solution we apply is to escape *all* non-ASCII characters. The
upside is conceptual simplicity; the downside is that non-ASCII
characters use 2-3x more bandwidth when transferring.

http://bugzilla.zimbra.com/show_bug.cgi?id=74240