Bug 74586
I found at least 2 bugs:
1. sample use-case: in conv view, click on a conv that has a few messages, some read. Then click on the header (subject) to expand all. Then do the quick reply - the new message body does not appear in the open conv.
solution:
this is due to the msgs collection being re-created in
ZmConv.prototype._handleResponseLoadMsgs, thus losing the listeners.  the solution is in ZmConv.prototype._handleResponseLoadMsgs, instead of re-creating a new ZmMailList, re-use the existing one, just remove all the items.
2.  Timing issue. sample use-case: click on a conv in the conv list view. then double click on it to open it in a standalone tab. In the standalone
tab, quick reply. The new message does not appear in the standalone view. (though usually a new element is added there, with footer, just no message body).
solution:
The cause is incomplete handling of the "loading" case for a msg. (started in bug 60257)
The reason for the race condition is that there are multiple views trying to load the same message, e.g. the conv list view and the convview2 "main" (reading pane), and in the case a standalone conv tab, that view too. Since ZmMailMsg.prototype.load was calling the callback in the
case the message is loading, we got partial results.
One more thing - on the way to the solution, I also thought it makes sense to remove listeners when closing (via the "close" button) a standalone conv view. However, this is not complete -  looks to me like the tab close button does not initiate any real cleanup (other than deactivate that doesn't do much, but in this case it's not just deactivate, it has to be disposed).

http://bugzilla.zimbra.com/show_bug.cgi?id=74586