bug 73702 - Invite and share buttons broken for single message in conv. view
The ZmMailMsgView created for an expanded single message in the
conversation view never received the invite reply listener. In order
to address this, we stash any given listeners on the object and apply
them when creating new views. Share and subscribe listeners used a
similar code path, and at least share listeners were similarly broken.

http://bugzilla.zimbra.com/show_bug.cgi?id=73702