Bug 75752
Js error "n.list is undefined" while opening a message in separate window.
once we expand the conv via the header, ZmConv.prototype._handleResponseLoadMsgs is called and had a bug in the call to ZmMailMsg.createFromDom, as the list should be passed as an attribute of args, and not standalone.

http://bugzilla.zimbra.com/show_bug.cgi?id=75752