Bug:73645
Tinymce's showmenu and hidemenu method is overridden. When menu is opened, notification will be sent to ZmAdvancedHtmlEditor method for listening mousedown event outside the editor. When menu is closed the outside event listener registered will be cleaned up.

Tinymce is having default event notifier as "onShowMenu" and "onHideMenu" methods. But that will be instance specific. The fix done here is in prototype level.  

http://bugzilla.zimbra.com/show_bug.cgi?id=73645