Bug 75260
Multiple issues with bubble and misalignment in message reading pane
Added the ZPropertySheet to the table, for the padding to be correct. Then long story short - I refactored the templates to do the same thing for the "sent by" for both message and conv view - extracted a sub-template that uses the conv approach of using spans for the different parts (sent-by, obo and bwo).   The result is common code for
both places, that's simpler and cleaner.

http://bugzilla.zimbra.com/show_bug.cgi?id=75260