bug 74609 - IE: GAL members of group created during previous session absent
When initially loading the Address Book app, ZmContactList issues a
REST request to obtain all contacts. If any of these contacts are
groups containing members not otherwise seen -- such as GAL entries --
ZmContactList._handleReponseLoad() reloads them in a dereferencing manner.

Unlike other browsers, IE caches the contact REST request beyond
session boundaries. As a result, the initial load of all contacts does
not include the recently created group, which thus isn't
dereferenced. The end result is that we never obtain a contact
corresponding to the GAL members. Then, when displaying the group, we
hit TODOs in ZmContact.getGroupMembers[Obj] and silently drop the
member.

In order to suppress IE's cache, we append the session ID to contact
REST URL.

http://bugzilla.zimbra.com/show_bug.cgi?id=74609