Bug 75921
Replying/Forwarding to message from conversation view doesn't select recent received message from conversation
The issue i solved (one of a couple but solves the issue right now) - when clicking on "expand all" (the conv header) we call a GetConvRequest - this one returns the msgs in creation order (older first). However SearchConvRequest returns them newer first and that is the order we use elsewhere. ZmMailList.prototype.getFirstHit is called when the user clicks the "reply" (or reply all) from the toolbar. However we lose the inHitList info when doing the GetConvRequest (it does not return the "cm" element), so it defaults to the first one. But the first one was the oldest instead of newest.

http://bugzilla.zimbra.com/show_bug.cgi?id=75921