Bug 74169
Js error (params.folder is undefined) while undoing deleted conversation from search result
at least until we fix bug 74446 (if possible to fix), I do not allow "undo" of conv moves if the current search does not have a folderID. (as the undo falls back to it in the case the convs are not fully loaded, as it is the _fromFolderId. Of course we could check if all the convs moved are fully loaded and only if not, prevent the undo, but that would be over-complicated and more inconsistent from user perspective).

http://bugzilla.zimbra.com/show_bug.cgi?id=74169