Bug:  76142 - Separator line missing and body content showing together for forwarded invite message
This problem was exposed due to recent fixes. Earlier while forwarding an appointment both the html and plain text content was forwarded. Now if the compose option is text then only text/plain message is sent. When the server get's only text/plain it add's a text/html content by surrounding the plain text with a pre tag(I think server does this for requirement by Outlook clients). There were couple of issues with this. First the truncateBodyContent did not receive the information if the content was html or plain text. The second problem was that the content was just removed till the notes separator marker. This also removed the start of the html tags(in this case the pre tag in html content). Fixed the code to preserve the html tags if the content type is html.

http://bugzilla.zimbra.com/show_bug.cgi?id=76142