bug:75054
Enablelog has no affect for now.
Verbose and Loglevel work as below.
We can change loglevel to either Info,Debug or Trace.
If Verbose is on and loglevel is info, we make loglevel as debug to give more detail logging.
verbose  been on/off has no affect on loglevel.
this discrepancy is because UI has little scope of changing the screens.

http://bugzilla.zimbra.com/show_bug.cgi?id=75054