Bug:76466
If an image is pasted in html editor, it will have an dfsrc attribute indicating it as an attachment. If the same image is updated to some external image url, then dfsrc attribute has to be removed. Otherwise the image url is updated back to the attachment url after saving.

Introduced dummy attribute for removing dfsrc attribute if it is an external image.

http://bugzilla.zimbra.com/show_bug.cgi?id=76466