bug: 76323
JaxbProvisioning Unit test failures

Root cause was that tests were relying on fact that sending an auth request to an account would create the associated mailbox.  Turns out that was actually a bug!  Now do a GetMailboxRequest to cause mailbox creation instead.

Some WSDL tests were failing due to the same root cause, so fixed those in the same way.

Another Volume related WSDL test was making an erroneous assumption that compressionThreshold would be at least 256 - relaxed that.

The original fix in TestJaxbProvisioning uncovered another issue:
java.lang.AssertionError
        at com.zimbra.cs.index.LuceneIndex.deleteIndex(LuceneIndex.java:175)
        at com.zimbra.cs.mailbox.MailboxIndex.deleteIndex(MailboxIndex.java:308)
        at com.zimbra.cs.mailbox.Mailbox.deleteMailbox(Mailbox.java:2050)
        at com.zimbra.cs.mailbox.Mailbox.deleteMailbox(Mailbox.java:1977)
        at com.zimbra.cs.service.admin.DeleteMailbox.handle(DeleteMailbox.java:77)
        at com.zimbra.soap.SoapEngine.dispatchRequest(SoapEngine.java:421)

The problem is that there are modifications to the index (reindex etc) going on in parallel and deleteMailbox makes no attempt to handle that.
Now deleteMailbox() sets a flag to indicate that delete is pending and then waits until all writers have finished (signalled via a Condition) before doing the actual delete.
The flag is used by various potential writers to stop trying to update the index.  Also, any attempts to open more writers will fail with a custom IOException.

https://reviewboard.eng.vmware.com/r/361901/

http://bugzilla.zimbra.com/show_bug.cgi?id=76323