Bug 76888
group/dl edit view code cleanup and fix large icons
Fixed the unit tests. One problem was due to some older fix I did to refactor the ZmGroupView to reuse code from ZmContactPicker. The others are due to my changes in this here bug, the dedupe changes, etc.
Also thanks to doing the tests, found some bad documentation and a no longer used field so fixed that too. Also decided to double-check the type as well as the value when de-dup just in case multiple types could have same id.
Also just found a bug in the dedup of inline items if I repeat items more than once in the inline box.  And this made me find a bug in AjxUtil.dedup (bug 74561)

http://bugzilla.zimbra.com/show_bug.cgi?id=76888