Bug 76990
Script error when pressing Search button on search page
I can't repro it, but maybe it's a combination of a similar thing to bug 73538 (where the type is passed as scalar, but here it's not the filter case, so I don't know in what case it is), and then that fix (and also some other bugs) do not work well with the fix to bug 69423 since there it saves the previous types as passed in the params, and not the ones we fix to be vector, so they might still be scalar).
I do believe this fix should fix it, but I wish I was able to repro first.

http://bugzilla.zimbra.com/show_bug.cgi?id=76990