Bug 77118
"conversation moved to trash" toast fades out right away if next conv is unread
We do not support undoing a read/unread action (see ZmActionStack.prototype.logAction - action would always be null for read/unread as those ops are not supported there.)
In any case, this fix is good in general - if the action does not log in the action stack, do not dismiss the current toast. Very simple.

http://bugzilla.zimbra.com/show_bug.cgi?id=77118