bug: 77071
Unmarshal JSON containing KeyValuePairs (_attrs) into JAXB objects regardless of what element/attribute name is used in the XML representation of the pairs.

An intermediate step in the unmarshaling uses JSONElement structure which defaults the element/attribute name to "a"/"n" (as the JSON does not contain this information).
The fix is to utilize the additional information available in the JAXB objects to fixup this structure at a later stage.
Fix is also planned for Octopus - see Jira HS-8952

https://reviewboard.eng.vmware.com/r/372631/

http://bugzilla.zimbra.com/show_bug.cgi?id=77071