Integrate change 396684 from IRONMAIDEN to main
-------------
Bug: 77019 - Text is quoted twice in reply
Revert the fix for 76636, which added a newly fetched alternative part to the
msg's list of body parts, making the msg think it has multiple body parts when
it just has a single alternative one. The better fix is to check if a part has a
multi/alt ancestor, rather than just checking its parent. The MIME for 76636
has a multi/alt with text/plain and multi/relative children (the latter has the HTML
part), so the HTML part wasn't recognizing that it was part of a multi/alt.

http://bugzilla.zimbra.com/show_bug.cgi?id=77019