Bug 77206
Sending link to item in shared briefcase links to wrong user's briefcase
Well, for a remote item we need to get the restUrl correctly.  First of all, for a shared briefcase, the restUrl is already set, so definitely use it if it's a shared briefcase instead of the calculation in ZmOrganizer.prototype.getRestUrl that only works for local. Also, for a sub-folder of shared, we had special problem, that restUrl is NOT pre-set (not sure why, perhaps we should fix that but not now), so for that we need to make sure to call this._generateRestUrl which seems to work for remote (shared) subfolder.  This fix seems safe since it only affects remote folders.

http://bugzilla.zimbra.com/show_bug.cgi?id=77206