bug: 75104
Is this being implemented as expected?  When I delete the last msg. from a
conversation (ie., it's now in Trash), nonetheless that msg. is still appearing
in the snippet & message preview.  Which is exactly what I don't want.  I
thought I had understood that Trash and Junk msgs. in a conversation would not
show up in the snippet & preview ?


made changes to Remove trash and junk msgs from snippet

http://bugzilla.zimbra.com/show_bug.cgi?id=75104