Bug: 77504
Selenium Harness: 

FormApptNew.java: add small delay after clicking send.  The client seems slower than the harness for this action - the queue is being checked before the CreateApptRequest is being sent.  Maybe the busy overlay isn't being applied?

CreateMeeting.java: for the basic selenium test case for meeting requests, verify that the attendee and optional attendee receive a meeting invitation and that the appointmnet is placed in the calendar

http://bugzilla.zimbra.com/show_bug.cgi?id=77504