bug: 77304
SOAP error response is in XML if there is a problem parsing the XML, even if the first part of the XML is understandable and requests a JSON response.

If the XML for a Soap Request was bad, look at it to see if enough of it is valid to be able to determine the desired response protocol.
Use StAX parsing so that we can stop looking at the XML once we have got past the Envelope Header context.

Also detects when SOAP 1.1 is more appropriate that SOAP 1.2 for the response.

http://bugzilla.zimbra.com/show_bug.cgi?id=77304