Bug 77594
Preferences->Sign in options->Font doesn't seem work
Regression from bug 64559. It undid the Bug 56316 where I removed the @FontFamily-default@
from the common.css generic tags, so it can be chosen dynamically by the user font selection, setting the user_font_modern/wide/etc class to the body.
So this adds the class to the body of the html client as well. Had to move the user_font_modern/etc class definitions from zm.css to common.css as zm.css is not loaded on the html client.

http://bugzilla.zimbra.com/show_bug.cgi?id=77594