Integrate change 407248 from IRONMAIDEN to main
-------------
Bug: 76820 - moving message to different folder causes conversation to disappear
We were looking at the wrong search object to figure out if the search is
matchable and whether the conv still matches the current search after the
message move. We need to look at the search that's tied to the search tab,
and we were looking at the search backing the conv list in the Mail tab.

http://bugzilla.zimbra.com/show_bug.cgi?id=76820