Bug 77886
Use tag bubbles in contact views and refactor common code
refactored to use a helper static class to deal with most of the tags issues (html to create the bubbles, change listeners, etc). Refactored out some duplicate or similar code, and in the process of course updated the contacts views (view/edit) to use the tag bubbles instead of the old school ugly links. Now everything (mail, tasks, contacts) uses tag bubbles in consistent way and using the same main code.
Also found a bug - it was not updating the item's tags in case of rename for lists that were not the current view. Fixed that in ZmList.prototype._tagTreeChangeListener.

http://bugzilla.zimbra.com/show_bug.cgi?id=77886