Integrate change 407352 from IRONMAIDEN to main
-------------
Bug: 76674 - Conv view: changing display format can result in bad view
This was an async problem. While rendering a message we may need to make
a server request to get the alternative part (when switching between showing
mail as HTML or text), so we need to use a callback to display the footer after
the message body has been rendered. Previously, the footer was getting
rendered during the request.

http://bugzilla.zimbra.com/show_bug.cgi?id=76674