Bug 77742
Include shared items search preference not taken into account
It's harder and harder to really lazy load the search menu - and we load it
anyway on switching tabs. And I think it's not such a heavy process. So in this
change, I'm kinda reverting what bug 20438 fix was doing. It was fixing bug
17090 change, but now it's not enough as we need to know things earlier. For
now I think we should do this on startup too. The price of generating this menu
on load is probably so minor compared to other much heavier things, that should
not make any difference.

http://bugzilla.zimbra.com/show_bug.cgi?id=77742