Bug 77846
Blank space before character string not modified after editing in Address Book
This is what I think is the real fix. Trim the value inside DwtInputField.prototype.getValue. I do not see a reason EVER that we would want an input field with leading or trailing spaces kept as part of the value.
Instead of trimming (or forgetting to trim, or going through hoops to trim) the spaces in a lot of places, this works by doing it in one small function.
Also - this immediately appears to the user on unfocus from a frield - the leading spaces (if any) disappear... and if that's the only change in the form - the form is not considered dirty (and it is not), saving annoying dialogs when they try to leave the edit page without any change.

http://bugzilla.zimbra.com/show_bug.cgi?id=77846