Bug 77846
Blank space before character string not modified after editing in Address Book
Extend previous fix to allow suppression of trimming for special cases, and implement it for the only special case we found so far - passwords (in particular in the external account setup). We allow leading/trailing spaces in passwords.

http://bugzilla.zimbra.com/show_bug.cgi?id=77846