bug: 77752
Shared addressbook not alphabetical for non English locales

For localised languages, we internally substitute SortBy.NAME_LOCALIZED_ASC for SortBy.NAME_ASC
SortBy doesn't have a "name" for NAME_LOCALIZED_ASC intended for use in SOAP requests, however it was being used in the proxied SearchRequest.  The SearchRequest handler wasn't recognising it and substituted dateDesc sorting instead.  Fixed this by using "nameAsc" in the proxied request.

The results of the proxied request were being placed in an Array of ProxiedHits - these are expected to be sorted by A_SORT_FIELD - but contacts are sorted by A_FILE_AS_STR.  Introduced ProxiedContactHit to address this.

Seems crazy to sort the proxied hits again, so, introduced the isPreSorted() method for the ZimbraQueryResults interface and use that to skip the re-sort.  In theory this means that ProxiedContactHit isn't strictly needed as the second sort no longer happens but feel more comfortable with it left in.

https://reviewboard.eng.vmware.com/r/390764/

http://bugzilla.zimbra.com/show_bug.cgi?id=77752