Integrate change 411002 from zimbra to zimbra
-------------
Bug 77920 - Domain administrators unable to create aliases or distribution lists
The access to alias or DL is determined by looking up access for each of the domain. The domain list is available only if Domain View rights are granted. If the Domain View rights are not granted then the rights for the current domain should be checked.

http://bugzilla.zimbra.com/show_bug.cgi?id=77920