Bug 78058
While click on save with the valid signature name and  empty signature body It doesn't shows toast message as "Signature value is empty. It's required."
Fix to the bug. associate the "autoAdded" attribute with each signature that was auto added by the app. (this signatures are disposed if the signature is empty, and thus are legal that way, unlike any other signature that was already saved, or added explicitly by the user).

http://bugzilla.zimbra.com/show_bug.cgi?id=78058