bug: 77604
When a user has been configured to change their password on next login, the resulting proxied ChangePasswordRequest was failing because the account was specified in the context but no authentication token was supplied.
The server handler rejects a context which has account information but no authentication info - see ZimbraSoapContext constructor.

https://reviewboard.eng.vmware.com/r/393466/

http://bugzilla.zimbra.com/show_bug.cgi?id=77604