Bug 78223
columns do not line up with reading pane on bottom
fix my previous fix - calling reset was not a good idea since it sets _rendered to false,and in ZmListView.prototype._getItemsNeeded it returns nothing if _rendered is false - thus pagination does not happen (on scroll down).
This fix is better - override removeAll in ZmMailListView to call clearGroupSections.

http://bugzilla.zimbra.com/show_bug.cgi?id=78223