Integrate change 413397 from zimbra to zimbra
-------------
Bug 77477 - View Mail missing from open source upgrade
The View Mail code was extracted from the admin console and created as an extension. This extension got created in the ZimbraNetwork due to which the FOSS users were not able to use View Mail. Moved the View Mail extension from ZimbraNetwork to FOSS ZimbraAdminExt and updated the build scripts.

http://bugzilla.zimbra.com/show_bug.cgi?id=77477