bug: 78269
LmtpClient's construction of data for the DATA command failed to insert an additional period at the beginning of lines of mail text that start with a period if the second character of the line happens to also be a period. SMTP RFC 5321#section-4.5.2 Transparency requires that ALL lines starting with a period should have an additional period added (LMTP uses the same mechanism)

As a result, lines in messages imported using the lmtpinject tool were missing the first "." from any line starting with "..".

With a fix only in LmtpClient, RunUnitTests test TestSmtpClient.testTransparency() started failing.  This is because SmtpToLmtp also uses LmtpClient and relied in its incorrect behavior.
The code that read data from SMTP didn't remove the first "." of lines in DATA to re-assemble the original text, relying on LmtpClient's behavior of not adding an extra ".".  Fixed this too.

https://reviewboard.eng.vmware.com/r/395385/

http://bugzilla.zimbra.com/show_bug.cgi?id=78269