bug: 77786
When formatting a contact as a VCARD, some properties (e.g., N and ADR) comprise multiple fields delimited by a SEMICOLON character (U+003B).  Therefore, a SEMICOLON in a field of such a "compound" property MUST be escaped with a BACKSLASH character.

VCard "formatted" field encapsulated to need to use getFormatted()
https://reviewboard.eng.vmware.com/r/396402/

http://bugzilla.zimbra.com/show_bug.cgi?id=77786