Bug:78051 - Intermittently compose send, save & attach button remains disabled while adding/removing attachment or at auto saving draft
Most of the changes are merged from following check-in, that is done in IRONMAIDEN.  

Change 411754 by rottikunta@rottikunta-IRONMAIDEN on 2012/10/24 
        13:37:25

        Bug: 78050 - Intermittently compose send, save & attach button 
        remains disabled while adding/removing attachment or at auto saving 
        draft

        The issue is due to missing context while running _resetUpload 
        function. i.e When user cancel new message while upload is in 
        progress, view may not be ZmComposeView. Since _resetUpload depends 
        on currentView, it throws JS errors. 

        Similarly user can switch the views while upload is in progress. i.e 
        When user goes to calendar tab when upload in progress, subsequent 
        _resetUpload call throws JS errors. Resolved most of the context 
        issues while uploading the attachment

Added support for canceling a specific file (that is being uploading or already uploaded), when user selects multiple files for uploading.  Set temporary bubble div id as attachment id and it is used for checking whether user deleted the attachment or not. Skipped this check for last attachment bubble as we save the message as draft immediately after last attachment is uploaded.

http://bugzilla.zimbra.com/show_bug.cgi?id=78051