Bug 77217
from search results page, unable to move message in trash to inbox
If we know all items moved (every message within every conversation selected by the user) are inside special folders (any of Sent/Drafts/Trash/Junk) - allow the move without any tcon.
This is the general solution that takes care of the case described in the bug description - a single-message conv that is in trash.
I also refactored and cleaned up (giving more meaningful variable names) the previous getTcon code. (that is still called in the general case not all items are special).

http://bugzilla.zimbra.com/show_bug.cgi?id=77217