Bug: none
Selenium Harness: dev environment changes

Don't use postqueue to check for message delivery.  Instead, use a sleep delay (configurable in config.properties).  

Also, remove the old "DevEnvironment" setting.  This was intended to be a single point that combined all the settings required to work in the dev env.  That requirement is starting to go away since dev is attempting to use a WDC client.  Instead of a single configuration point, try to configure each interaction point individually in the code.