Bug 78882
Sometimes the entire screen is shifted up and corrupting
skin_container_app_chooser element was defaulting to height "100%" (from some generic style). this caused it to be the height of it's container, un-necessarily, and thus allowing the scroll of its sibling TR height (since
for some reason when setting height 100%, in the static [i think] positioning of the container case - the siblings height is not deducted when computing height 100%). The sibling is the Zimbra header row so that is why the container scrolls by that height. It's hard to explain but anyway this seems to solve the problem.

http://bugzilla.zimbra.com/show_bug.cgi?id=78882