Bug 78980
JS error when doing delete with messages between group divider
Fix is to not add the separator (group divider) to the selected items. (I could have simply added some null check in the place of the error, but that would have only solved one symptom of the bigger problem, I believe).

http://bugzilla.zimbra.com/show_bug.cgi?id=78980