Bug 77538 Unable to delete email when its opened for viewing
I don't like fixing bugs without being able to repro, but this fix seems to not break anything (no reason it should) and would fix the exception this bug is talking about. (since the conditional code will no longer be executed in case of ZmMsgController).  Hopefully the same bug causing this exception won't reveal itself in a different way (since the code here won't stop the flow), but I can't tell for sure.

http://bugzilla.zimbra.com/show_bug.cgi?id=77538