Bug 51102
Name of first persona reverted in list when adding another
the bug only affects the list and in a weird way. When you click on the "New Persona 1" row it goes back to the name you entered "aaaa" for example. It also saves correctly.
This fixes the issue - The _currentAccount's name was not updated on the change.

http://bugzilla.zimbra.com/show_bug.cgi?id=51102