Bug 79414
Some contact groups corrupted after upgrade to 8.0.2
The problem is in ZmContact.prototype.getAllGroupMembers we call ZmContactsHelper._wrapInlineContact with the value. I assumed incorrectly the value was just an email address, but it could be (at least legacy data, not sure about newer data) in the format "John Doe <x.com>". The best place to support this is in ZmContactsHelper._wrapInlineContact itself so that's where I fixed it - by calling AjxEmailAddress.parse on the string.

http://bugzilla.zimbra.com/show_bug.cgi?id=79414