Bug:79399 Integration change from Bug:70482
Appointment reminder settings do not seem consistent. The reminder Controller housekeeping action never invoked if the default reminder time was zero. This is an artifact from an earlier implementation, which used the default reminder period to calculate an end time for searching for reminders to display.  The code is now unused (and is disabling the housekeeping when zero), so I removed it.

http://bugzilla.zimbra.com/show_bug.cgi?id=79399