Bug 74742
Limit the ability to resize columns so that you can't make them so narrow the resize bar disappears
The problem was that the TD that contains the label (inner TD inside the header TD) , did not have a limit to the size, so for example "Folder" always took the whole width. Therefor the TD after it, for the sash - did not show if the container was not big enough so the label overflowed thus pushing the sash to hide too.
The solution is to have the the attribute table-layout:fixed on the inner table in each column header - that way I set the sash to fixed size (6px), the arrow to 8px and the label stays without a size so it just takes as much as remains (as a filler).

http://bugzilla.zimbra.com/show_bug.cgi?id=74742