Bug 79524
Can't delete multiple appointments from calendar list view using Delete toolbar button and context menu
This is a regression from my fix to Bug 78398 that included a change of constant (enum-like) mode values from numbers to meaningful names  (seemingly innocent change -  just for better debugging).   And for some reason the code for bug 19338 it was doing parseInt to those numbers, that has no effect on numbers but gets a NaN for Strings. Doh.

http://bugzilla.zimbra.com/show_bug.cgi?id=79524