Bug 79525
"Sort by subject" missing while using reading pane on the right
Indeed regression from bug 74742.
introduced a param isColumnHeaderTableFixed (not the best name, but it's kind of not a case that's easily characterized in one or 2 words) that is passed to DwtListView.prototype.createHeaderHtml (conveniently it can be passed as needed from subclasses of DwtListView as a lot of them already override DwtListView.prototype.createHeaderHtml and call it from the overriding method).
Currently I set this param to true for mail (only if multi-column case, i.e. view pane off or at bottom, so this will solve this case of view pane on right), and also on calendar list view.

http://bugzilla.zimbra.com/show_bug.cgi?id=79525