Bug 79681
IE : Background color for mail not set as tag color
the problem reported in the bug was due to single quotes inside the CSS (around the color), and the caller used single quote to wrap the style attribute thus the first single quote would terminate the style prematurely.
I decided the simplest and maybe most error proof solution was to remove the single quotes around the colors - it worked so why not.  That way callers can wrap the style in double quotes as well as single quotes (as long as no new style introduced in the method Dwt.createLinearGradientInfo includes either).
There was an additional bug - it was still not working in the case of preview pane on the right - Josh then figured it was due to size not calculated correctly and suggested adding a zoom: 1 style - this did the trick for that.
We also noticed the selection highlighting was not overriding the tag gradient color in IE - so fixed that too - had to override the filter.

http://bugzilla.zimbra.com/show_bug.cgi?id=79681