bug:79208,79634 Introduce new ldap attributes zimbraHttpConnectorMaxIdleTimeMillis and zimbraHttpThreadPoolMaxIdleTimeMillis that can be configured on globalConfig or server.
zimbraHttpConnectorMaxIdleTimeMillis - Maximum Idle time in milli seconds for a connection. This is applied when waiting for a new request to be received on a connection; when reading the headers and content of a request; when writing the headers and content of a response. Default value 60000

zimbraHttpThreadPoolMaxIdleTimeMillis - The maximum thread idle time in milli seconds. Threads that are idle for longer than this period may be stopped. Default value 10000

https://reviewboard.eng.vmware.com/r/421035/

http://bugzilla.zimbra.com/show_bug.cgi?id=79208
http://bugzilla.zimbra.com/show_bug.cgi?id=79634